(refactor) Merge logic to render different workout types from editAMRAP to instead be one level higher, at the modifyWorkout component
